Game 클래스의 입력부를 Input 클래스로 분리하기
다른 분들의 코드를 보면서 별로 노력하지 않았다는 것을 느껴서 급하게 수정하였습니다.
