+++ README.md
@@ -76,3 +76,4 @@
 16. Race 클래스의 책임이 분명해지도록 클래스명을 리팩토링 하였습니다.
 17. Game 클래스의 인스턴스 변수를 private로 접근 제어자를 설정해주었습니다. (왜 뺴먹었지...)
 18. 소스가 있는 패키지의 위치를 이동하였습니다.
+19. Game Class에서 입력받는 부분을 Input Class로 분리하였습니다. // 싱글톤 패턴..?
